int/builder: organizes files.
